Bootstrap Confidence Intervals

This function resamples the data and lets the user choose the significance level α to obtain for the 1−α confidence intervals for β, and the number of bootstraps which by default is 20.

bootstrap_conf_intervals(X, y, alpha = 0.05, n_bootstraps = 20)

Arguments

alpha

scaler statistical significance value (default is 0.05)

n_bootstraps

codecharacter defines the number of times the bootstrap will re-sample the data (default is 20)

resp

y vector of length n.

pred

X matrix n number rows beta plus 1 number of columns with first column being entirely ones.

Value

A vector that contains the confidence intervals of the bootstrap

Author

Elena Gagliano

Helen Wu

Max Van Horn

Examples

set.seed(123)
x1 <- rnorm(100)
x2 <- rnorm(100)
x3 <- rnorm(100)
X <- cbind(1,x1, x2, x3) #1 for the intercept term
beta0 <- -1
beta1 <- 2
beta2 <- -0.5
beta3 <- 1.5
true_beta <- matrix(c(beta0, beta1, beta2, beta3), ncol = 1)
p <- 1/(1 + exp(-(X%*%true_beta)))
y <- rbinom(100,1,p)
B_initial <- solve(t(X)%*%X)%*%t(X)%*%y
bootstrap_conf_intervals(X, y)
#>               lower       upper
#> Intercept -1.314399 -0.08886915
#> x1         1.966590  3.43555256
#> x2        -1.378364 -0.09329511
#> x3         1.364147  2.62281628